requirements.txt
# RAG Question Answering System
# Full English Version for GitHub

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Vector & Retrieval
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
rank-bm25>=0.2.2

# Text Processing
jieba>=0.42.1
nltk>=3.8.1
stopwordsiso>=0.6.1
langchain-text-splitters>=0.0.1

# Document Parsing
PyMuPDF>=1.23.0
python-docx>=1.1.0
EbookLib>=0.18
beautifulsoup4>=4.12.0

# HTTP & Database
requests>=2.31.0
pymysql>=1.1.0

# Data & Utils
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6